+++ @generated/graphql/index.ts
@@ -1297,6 +1297,7 @@ export type CommertialOffer = {
   Installments: Maybe<Array<Installment>>
   ListPrice: Maybe<Scalars['Float']>
   Price: Maybe<Scalars['Float']>
+  discountHighlights: Maybe<Array<DiscountHighlight>>
 }
 
 export type DateQueryOperatorInput = {
@@ -1635,6 +1636,10 @@ export type DirectoryGroupConnectionSumArgs = {
 export type DirectorySortInput = {
   fields: InputMaybe<Array<InputMaybe<DirectoryFieldsEnum>>>
   order: InputMaybe<Array<InputMaybe<SortOrderEnum>>>
+}
+
+export type DiscountHighlight = {
+  name: Scalars['String']
 }
 
 export type File = Node & {
@@ -4156,6 +4161,7 @@ export type ProductSummary_ProductFragment = {
         Name: string
         PaymentSystemName: string
       }> | null
+      discountHighlights: Array<{ name: string }> | null
     } | null
   }> | null
 }
@@ -4528,6 +4534,7 @@ export type ProductsQueryQuery = {
                 Name: string
                 PaymentSystemName: string
               }> | null
+              discountHighlights: Array<{ name: string }> | null
             } | null
           }> | null
         }
@@ -4575,6 +4582,7 @@ export type SearchSuggestionsQueryQuery = {
               Name: string
               PaymentSystemName: string
             }> | null
+            discountHighlights: Array<{ name: string }> | null
           } | null
         }> | null
       }>