+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added- 
 - `Badge Price` on cart and search suggestion ([#50](https://github.com/vtex-sites/accenturebb.store/pull/50)
+- `Discount Name` on product card ([#49](https://github.com/vtex-sites/accenturebb.store/pull/49)
 - `PDP CMS page` on content-types as a new page ([#46](https://github.com/vtex-sites/accenturebb.store/pull/46)
 - `Product Card` add installment ([#42](https://github.com/vtex-sites/accenturebb.store/pull/42)
 - `react-simple-image-viewer` package to do zoom on PDP images ([#29](https://github.com/vtex-sites/accenturebb.store/pull/29)